Fix whitespace when unpacking IcoMoon folders
In the conformist fashion typical of modern web developers, IcoMoon have
decided to use spaces instead of tabs for their generated code. Not only
does this bloat the JSON's filesize pointlessly, it also risks causing a
mess of diff-noise in future commits.

This is only a temporary solution until IcoMoon can be dropped in favour
of a more streamlined and flexible process.